separate out into `encoder.js` and `decoder.js` files
For requiring individually with browserify to get smaller builds.
